Updated types.

+++ service/JavaService/src/main/java/io/eaglejs/ffxi/mapper/PlayerMapper.java
@@ -120,7 +120,7 @@ public class PlayerMapper {
         player.setTitle(document.getString("title"));
         player.setNationRank(document.getInteger("nationRank"));
         player.setStatus(document.getInteger("status"));
-        player.setLastOnline(document.getLong("lastOnline"));
+        player.setLastOnline(document.getInteger("lastOnline"));
         player.setGil(document.getInteger("gil"));
         player.setHpp(document.getInteger("hpp"));
         player.setMpp(document.getInteger("mpp"));
